src/requirements.txt
apache-beam==2.48.0
apache-flink==1.18.0
apache-flink-libraries==1.18.0
avro-python3==1.10.2
certifi==2024.8.30
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
fastavro==1.9.7
fasteners==0.19
find_libpython==0.4.0
grpcio==1.66.2
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
kafka-python==2.0.2
numpy==1.24.4
objsize==0.6.1
orjson==3.10.7
pandas==2.2.3
pemja==0.3.0
proto-plus==1.24.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zstandard==0.23.0
